+++ flink-streaming-connectors/flink-connector-akka/src/test/scala/org/apache/flink/connectors/akka/streaming/AkkaSinkTest.scala
@@ -1,0 +1,47 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.flink.connectors.akka.streaming
+
+import java.util.concurrent.TimeUnit
+
+import com.typesafe.config.ConfigFactory
+import org.apache.flink.runtime.client.JobExecutionException
+import org.apache.flink.streaming.api.scala._
+import org.apache.flink.streaming.util.StreamingMultipleProgramsTestBase
+
+import org.junit.Test
+
+class AkkaSinkTest extends StreamingMultipleProgramsTestBase {
+
+  implicit val timeout = akka.util.Timeout(10L, TimeUnit.SECONDS)
+
+
+  /**
+    * JobExecutionException wraps IllegalArgumentException in case of missing remote actor
+    */
+  @Test(expected = classOf[JobExecutionException])
+  def checkExceptionConfigSize(): Unit = {
+    val env = StreamExecutionEnvironment.getExecutionEnvironment
+
+    env.generateSequence(0, 20L).addSink(new AkkaSink[Long](path = "wrongPath"))
+
+    env.execute()
+  }
+}
+
+
